Refactor SelectBooks to encapsulate the booklist data structure
Change the name of BookListCreate and BookListCreate.test to Library
and Library.test respectively, as well as change the name of SelectBook
to Books. Delete the SelectBook test in the process as it is redundant.

Change both the ReadingList and Library getBooks() functions to return
instances of the Books class instead of their respective list of books.

Clean up validation in App to be more broad and fix bug introduction.

Update displayList function to properly display books from the Books
domain object instead of the list of books.